# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LSP UGM API Documentation",
        "description": "API Documentation for LSP UGM Application",
        "contact": {
            "name": "Administrator",
            "email": "admin@lsp-ugm.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "LSP UGM API Server"
        }
    ],
    "paths": {
...
        "/asesmen/mapa01/{id_asesi}": {
            "get": {
                "tags": [
                    "MAPA01"
                ],
                "summary": "Mendapatkan data formulir MAPA01 untuk asesi",
                "description": "Get MAPA01 data for an Asesi",
                "operationId": "798e85dc109e31c7929fba150925b6d0",
                "parameters": [
                    {
                        "name": "id_asesi",
                        "in": "path",
                        "description": "ID asesi",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data formulir MAPA01",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "general_info": {
                                                    "properties": {
                                                        "nama_asesi": {
                                                            "type": "string",
                                                            "example": "John Doe"
                                                        },
                                                        "nama_asesor": {
                                                            "type": "string",
                                                            "example": "Jane Smith"
                                                        },
                                                        "nama_tuk": {
                                                            "type": "string",
                                                            "example": "TUK JTTC"
                                                        },
                                                        "judul_skema": {
                                                            "type": "string",
                                                            "example": "Programmer"
                                                        },
                                                        "kode_skema": {
                                                            "type": "string",
                                                            "example": "SKM-001"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "rencana_asesmen": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id_uk": {
                                                                "type": "string",
                                                                "example": "UK-001"
                                                            },
                                                            "kode_uk": {
                                                                "type": "string",
                                                                "example": "UK001"
                                                            },
                                                            "nama_uk": {
                                                                "type": "string",
                                                                "example": "Mengembangkan Aplikasi Web"
                                                            },
                                                            "rencana": {
                                                                "type": "array",
                                                                "items": {
                                                                    "properties": {
                                                                        "elemen": {
                                                                            "type": "string",
                                                                            "example": "Menyiapkan kebutuhan"
                                                                        },
                                                                        "bukti_bukti": {
                                                                            "type": "string",
                                                                            "example": "Portofolio"
                                                                        },
                                                                        "jenis_bukti": {
                                                                            "type": "string",
                                                                            "example": "L"
                                                                        }
                                                                    },
                                                                    "type": "object"
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "mapa01": {
                                                    "properties": {
                                                        "pendekatan_asesmen_asesi": {
                                                            "type": "string",
                                                            "example": "Pendekatan asesmen yang digunakan"
                                                        },
                                                        "tujuan_asesmen": {
                                                            "type": "string",
                                                            "example": "Tujuan pelaksanaan asesmen"
                                                        },
                                                        "lingkungan": {
                                                            "type": "string",
                                                            "example": "Tempat kerja nyata"
                                                        },
                                                        "peluang_untuk_mengumpulkan_bukti": {
                                                            "type": "string",
                                                            "example": "Tersedia"
                                                        },
                                                        "hubungan_antara_standar_kompetensi": {
                                                            "type": "string",
                                                            "example": "Keterkaitan dengan standar industri"
                                                        },
                                                        "pelaksana_asesmen": {
                                                            "type": "string",
                                                            "example": "Tim asesor LSP UGM"
                                                        },
                                                        "pihak_yang_relevan_untuk_dikonfirmasi": {
                                                            "type": "string",
                                                            "example": "Atasan, rekan kerja, klien"
                                                        },
                                                        "tolak_ukur_asesmen": {
                                                            "type": "string",
                                                            "example": "Kriteria unjuk kerja dan pengetahuan pendukung"
                                                        },
                                                        "waktu_tanda_tangan_asesor": {
                                                            "type": "string",
                                                            "format": "date",
                                                            "example": "2025-05-15",
                                                            "nullable": true
                                                        },
                                                        "tanda_tangan_asesor": {
                                                            "type": "string",
                                                            "format": "date",
                                                            "example": "signature/ttd_asesor.png"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "record_exists": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Asesi tidak ditemukan atau belum memiliki rincian asesmen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Asesi tidak ditemukan"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "API_KEY",
                "in": "header"
            },
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "API_KEY",
                "in": "header"
            }
        }
    },
...
}